Client component directory structure refactor
