+++ client/src/components/project-card/index.js
@@ -12,7 +12,7 @@ import {
   Link,
 } from "@material-ui/core";
 
-import LinkTo from "../components/navigation/LinkTo";
+import LinkTo from "../navigation/LinkTo";
 
 moment.updateLocale("en", { relativeTime: { future: "%s to go" } });
